:construction: Add GoldsmithReport showIndicatorVisualisations
Introduced functionality to show multiple visualisations

+++ opensrp-goldsmith/src/main/java/org/smartregister/goldsmith/util/ReportingConstants.java
@@ -9,6 +9,6 @@ public class ReportingConstants {
     }
 
     public static class ProgressTargets {
-        public static int PREGNANCY_REGISTRATION_TARGET = 4; // Todo -> How can the supervisor set this?
+        public static float PREGNANCY_REGISTRATION_TARGET = 4; // Todo -> How can the supervisor set this?
     }
 }
